V4_refactor/requirements/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
jinja2>=3.1.2
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
openai>=1.0.0
anthropic>=0.5.0
langchain>=0.1.0
langchain_community>=0.0.10
langchain_openai>=0.0.2
qdrant-client>=1.7.0
chardet>=5.2.0
flask>=3.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
Flask-SQLAlchemy>=2.5
psycopg2-binary>=2.9
# Dependencies for local model usage (only needed if LLM_BACKEND != "azure" in config.py)
# torch - REMOVED: PyTorch dependencies
# transformers - REMOVED: PyTorch dependencies  
# accelerate - REMOVED: PyTorch dependencies 